Move & update packages commands for controlled lists
